telegram-bot/requirements.txt
aiogram==3.22.0
python-dotenv==1.0.1   
uvloop==0.21.0
groq==0.31.0
google-search-results==2.4.2
aiohttp==3.10.5